// packages/frontend/src/generated/types/RailWagonManifestResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { RailWagonMeasurementResponse } from './RailWagonMeasurementResponse.ts'
import type { RailWagonWeightResponse } from './RailWagonWeightResponse.ts'

/**
 * @description Response DTO for the `rail_wagon_manifest` entity.
 */
export interface RailWagonManifestResponse {
  /**
   * @type string
   */
  createdAt: string
  /**
   * @type string, uuid
   */
  createdBy: string
  /**
   * @type string
   */
  declaredDensity: string
  /**
   * @type string
   */
  declaredMass: string
  /**
   * @type string
   */
  declaredVolume: string
  /**
   * @type string,null
   */
  deletedAt?: string | null
  /**
   * @type string,null, uuid
   */
  deletedBy?: string | null
  /**
   * @type string, uuid
   */
  id: string
  /**
   * @type array,null
   */
  measurements?: RailWagonMeasurementResponse[] | null
  /**
   * @type string, uuid
   */
  originDbId: string
  /**
   * @type string, uuid
   */
  productId: string
  /**
   * @type string, uuid
   */
  railWaybillId: string
  /**
   * @type string
   */
  updatedAt: string
  /**
   * @type string, uuid
   */
  updatedBy: string
  /**
   * @type string
   */
  wagonNumber: string
  /**
   * @type array,null
   */
  weights?: RailWagonWeightResponse[] | null
}