// packages/frontend/src/generated/types/RailWagonMeasurementResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

/**
 * @description Response DTO for the `rail_wagon_measurement` entity.
 */
export interface RailWagonMeasurementResponse {
  /**
   * @type string
   */
  calculatedMass: string
  /**
   * @type string
   */
  createdAt: string
  /**
   * @type string, uuid
   */
  createdBy: string
  /**
   * @type string,null
   */
  deletedAt?: string | null
  /**
   * @type string,null, uuid
   */
  deletedBy?: string | null
  /**
   * @type string, uuid
   */
  id: string
  /**
   * @type string,null
   */
  labDensity?: string | null
  /**
   * @type string
   */
  measuredHeight: string
  /**
   * @type string, uuid
   */
  originDbId: string
  /**
   * @type string
   */
  updatedAt: string
  /**
   * @type string, uuid
   */
  updatedBy: string
  /**
   * @type string, uuid
   */
  wagonManifestId: string
}