// packages/frontend/src/generated/types/RailWagonWeightResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

/**
 * @description Response DTO for the `rail_wagon_weight` entity.
 */
export interface RailWagonWeightResponse {
  /**
   * @type string
   */
  createdAt: string
  /**
   * @type string, uuid
   */
  createdBy: string
  /**
   * @type string,null
   */
  deletedAt?: string | null
  /**
   * @type string,null, uuid
   */
  deletedBy?: string | null
  /**
   * @type string
   */
  grossWeight: string
  /**
   * @type string, uuid
   */
  id: string
  /**
   * @type string
   */
  netProductWeight: string
  /**
   * @type string, uuid
   */
  originDbId: string
  /**
   * @type string
   */
  tareWeight: string
  /**
   * @type string
   */
  updatedAt: string
  /**
   * @type string, uuid
   */
  updatedBy: string
  /**
   * @type string, uuid
   */
  wagonManifestId: string
}